BLAN-1114 - refactoring, moved sidebar components to a separate directory
