+++ src/views/application/index.vue
@@ -55,7 +55,7 @@ import { useI18n } from 'vue-i18n';
 
 import { v4 as uuid } from 'uuid';
 
-import Sidebar from './../../components/Sidebar.vue';
+import Sidebar from '../../components/Sidebar/Sidebar.vue';
 import MainHeader from './../../components/MainHeader.vue';
 import Notification from '../../components/Notification.vue';
 import VerifyEmailModal from '../../components/ModalWindows/UserCabinetModalWindows/VerifyEmailModal.vue';
